requirements.txt
aws-cdk.core
aws-cdk.aws_iam
aws-cdk.aws_sns
pylint